fix(ci): keep models.rs + endpoint_task.rs under 1000-line cap
Extract EndpointLifecycle (enum + compute + tests) to crates/rs-core/src/endpoint_lifecycle.rs
(re-exported from models.rs). Move rescue audit emits into rescue_audit::emit_activated/
emit_recovered helpers. models.rs 1095->932, endpoint_task.rs 1009->999.
