+++ crates/rs-core/src/lib.rs
@@ -1,6 +1,7 @@
 pub mod audit;
 pub mod config;
 pub mod db;
+pub mod endpoint_lifecycle;
 pub mod error;
 pub mod log_buffer;
 pub mod log_capture;
